Initial refactor of PageDefinition
