+++ src/components/Navigation/Pages/PageBrowse.js
@@ -1,21 +1,19 @@
-import React, { Component } from 'react'
+import React from 'react'
 import { bindActionCreators } from 'redux'
 import { connect } from 'react-redux'
-import PropTypes from 'prop-types'
 import { uiNavigation } from '../../../actions/ui'
 import { ROUTE_BROWSE } from '../../../utils/routingConstants'
+import AbstractPageDefinitions from '../../AbstractPageDefinitions'
 
-class PageBrowse extends Component {
+class PageBrowse extends AbstractPageDefinitions {
+  constructor(props) {
+    super(props)
+  }
+
   componentDidMount() {
     uiNavigation({ to: ROUTE_BROWSE })
   }
-
-  render() {
-    return <div />
-  }
 }
-
-PageBrowse.propTypes = {}
 
 function mapStateToProps(state) {
   return {
